fix: move data to root
